move legacy data creation methods legacy folder
