+++ tests/autorag/data/legacy/corpus/test_langchain.py
@@ -3,8 +3,8 @@ import tempfile
 import pytest
 from langchain_core.documents import Document
 
-from autorag.data.corpus import langchain_documents_to_parquet
-from tests.autorag.data.corpus.test_base import validate_corpus
+from autorag.data.legacy.corpus import langchain_documents_to_parquet
+from tests.autorag.data.legacy.corpus.test_base import validate_corpus
 
 
 @pytest.fixture
